TracRaw_lvlock: revert to int32 — the REMCLU bit (0x80000000) is real
Earlier commit 83595b8 changed the schema to int8 under the mistaken
belief that the high-bit values C++ ROOT was returning were "INT32_MIN
garbage from an RNTuple read bug" (because uproot on the same file
showed only 0/1). The diagnosis was backwards:

  - C++ ROOT (and PyROOT, which uses the same reader) faithfully reads
    the 4-value LVLOCK distribution {0, 1, 0x80000000, 0x80000001}.
  - uproot's RNTuple int32 decoder is what silently drops the high bit
    in some narrow-range vector<int32_t> fields.
  - 0x80000000 is *not* garbage. It is `SBIT1(LVLOCK(NVECP),32)` set by
    SKELANA's PSHREMCLU at skelana.car lines 2110-2113 to mark a
    Particle whose calorimeter energy has already been merged into a
    charged-track PFO (REMCLU overlap).

Truncating LVLOCK to int8 silently throws bit 32 away. The result is a
~7-10% population of Particles that should have been rejected for
REMCLU overlap but instead pass the `LVLOCK == 0` cut as 0. Including
them in event sums double-counts calorimeter energy — physics-breaking
in thrust, EEC, missing-pt.

Empirical confirmation (50-event Z->hadrons MC, same SDST):
  - int8 schema (my regression):  LVLOCK ∈ {0, 1}  only
  - int32 schema (this commit):   LVLOCK ∈ {0, 1, 0x80000000, 0x80000001}
                                 matching legacy Part_lock distribution.

Closure vs Jingyu's full pipeline (his nanoaod_writer + his
analysis_thrust.py) on the same SDST: 49/49 events bit-identical with
no "bug emulation" hack, mean Δ(T) = 0.00000000 ± 0.00000000.

Reference: forDickyChant/LVLOCK_bug.md (Jingyu Zhang, 2026-05-13) for
the SKELANA Fortran walkthrough.

+++ delphi-raw-nanoaod/include/raw_nanoaod_writer.hpp
@@ -233,11 +233,18 @@ private:
     std::shared_ptr<std::vector<std::int16_t>>            TracRaw_ndfVD_;         // Q(LMAIN+27)
     std::shared_ptr<std::vector<float>>                   TracRaw_chi2VDHits_;    // Q(LMAIN+18)
     std::shared_ptr<std::vector<std::int8_t>>             TracRaw_charge_;        // sign of Q(LMAIN+8)
-    // int8 (not int32) avoids a ROOT 6.38 RNTuple read-decoding bug specific
-    // to small-range vector<int32_t> fields (the C++ reader returns INT32_MIN
-    // garbage past element ~25 in some events; uproot reads the same file
-    // correctly). LVLOCK fits easily in 8 bits (values are 0/1/few-bit-flag).
-    std::shared_ptr<std::vector<std::int8_t>>             TracRaw_lvlock_;        // sk::LVLOCK at the matching VECP entry; 0 = passes IFLSTR=11/IFLCUT=3
+    // int32 to preserve the FULL SKELANA LVLOCK bitmask. The high bit
+    // (bit 32, hex 0x80000000) is set by PSHREMCLU when a Particle is
+    // part of a REMCLU calorimeter cluster whose energy has already been
+    // merged into a charged-track PFO — keeping these Parts in event
+    // sums causes calorimeter-energy double counting in thrust, EEC,
+    // missing-pt, etc. Truncating to int8 silently drops this bit, so
+    // the previous int8 schema (commit 83595b8) was a regression. The
+    // legacy `Part_lock` branch in nanoaod_writer.cpp is int32 for the
+    // same reason. Convention: an analysis selects a particle iff
+    // LVLOCK == 0 exactly (all 32 bits zero) — see skelana.car lines
+    // 909, 969 in the SKELANA Fortran source.
+    std::shared_ptr<std::vector<std::int32_t>>            TracRaw_lvlock_;        // sk::LVLOCK at the matching VECP entry; 0 = passes IFLSTR=11/IFLCUT=3, bit 32 = REMCLU overlap
     // BS- and PV-corrected impact parameters at sk::QTRAC(38..40, vecp_i).
     // Filled by SKELANA's PSCBHP (BS fit) + PV fit at analysis time and
     // stored back into the QTRAC bank's extension slots. Same values
